Added primitive tests for JSON each row
