--- client-v2/src/test/java/com/clickhouse/client/api/data_formats/GsonJSONEachRowFormatReaderTest.java
@@ -1,8 +1,0 @@
-package com.clickhouse.client.api.data_formats;
-
-public class GsonJSONEachRowFormatReaderTest extends AbstractJSONEachRowFormatReaderTest {
-    @Override
-    protected String getProcessor() {
-        return "GSON";
-    }
-}
