+++ client-v2/src/test/java/com/clickhouse/client/api/data_formats/GsonJSONEachRowFormatReaderTests.java
@@ -1,0 +1,11 @@
+package com.clickhouse.client.api.data_formats;
+
+import org.testng.annotations.Test;
+
+@Test(groups = {"integration"})
+public class GsonJSONEachRowFormatReaderTests extends AbstractJSONEachRowFormatReaderTests {
+    @Override
+    protected String getProcessor() {
+        return "GSON";
+    }
+}
